ТЗ-13 итерация 2

+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -3,7 +3,12 @@ package ru.practicum.shareit.item;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exceptions.ObjectNotFoundException;
+import ru.practicum.shareit.exceptions.ValidateException;
+import ru.practicum.shareit.user.UserService;
 
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 @Service
@@ -12,47 +17,81 @@ import java.util.List;
 public class ItemService {
 
     private final ItemRepository itemRepository;
+    private final UserService userService;
+    private final ItemMapper itemMapper;
 
-    public List<Item> findAllByOwnerId(Integer userId) {
+    public List<ItemDto> findAllByOwnerId(Integer userId) {
 
         log.info("I received a request to search for all the items added by the user to the id " + userId);
 
-        return itemRepository.findAllByOwnerId(userId);
+        List<Item> itemsList = itemRepository.findAllByOwnerId(userId);
+        List<ItemDto> dtoList = new ArrayList<>();
+
+        for (Item item : itemsList) {
+            dtoList.add(itemMapper.objectToDto(item));
+        }
+
+        return dtoList;
     }
 
-    public Item findById(int itemId) {
+    public ItemDto findById(int itemId) {
 
         log.info("Searching for a item with an id " + itemId);
 
-        return itemRepository.findById(itemId);
+        return itemMapper.objectToDto(itemRepository.findById(itemId));
     }
 
-    public Item create(Item item) {
+    public ItemDto create(ItemDto dto, Integer userId) {
 
-        Item newItem = itemRepository.create(item);
-        log.info("I received a request to create a item " + newItem);
+        if (userService.findById(userId) == null) {
+            throw new ObjectNotFoundException("There is no user with this id");
+        }
 
-        return newItem;
+        Item item = itemMapper.dtoToObject(dto, userId);
+        ItemDto itemDto = itemMapper.objectToDto(itemRepository.create(item));
+
+        log.info("I received a request to create a item " + itemDto);
+
+        return itemDto;
     }
 
-    public Item change(Item item) {
+    public ItemDto update(ItemDto dto, Integer userId) {
 
-        Item newItem = itemRepository.change(item);
-        log.info("I received a request to update a item\n" + "Old item " + item +  "\nNew item " + newItem);
+        if (userService.findById(userId) == null) {
+            throw new ObjectNotFoundException("There is no user with this id");
+        }
 
-        return newItem;
+        Item item = itemMapper.dtoToObject(dto, userId);
+        ItemDto itemDto = itemMapper.objectToDto(itemRepository.update(item));
+
+        log.info("I received a request to update a item\n" + "Old item " + item + "\nNew item " + itemDto);
+
+        return itemDto;
     }
 
-    public void deleteItem(int itemId) {
+    public void deleteItem(int itemId, Integer userId) {
 
-        itemRepository.deleteItem(itemId);
+        if (findById(itemId).getOwnerId().equals(userId)) {
+            itemRepository.deleteItem(itemId);
+        } else {
+            throw new ValidateException("Only its owner can delete an item");
+        }
+
         log.info("I received a request to delete a item with an id " + itemId);
     }
 
-    public List<Item> search(String text) {
+    public List<ItemDto> search(String text) {
 
         log.info("I received a request to search for things whose name or description contains text: " + text);
 
-        return itemRepository.search(text);
+        List<ItemDto> dtoList = new ArrayList<>();
+        List<Item> itemsList = itemRepository.search(text);
+
+        for (Item item : itemsList) {
+            dtoList.add(itemMapper.objectToDto(item));
+        }
+        dtoList.sort(Comparator.comparing(ItemDto::getId));
+
+        return dtoList;
     }
 }
